+++ src/Fable.Cli/Util.fs
@@ -93,17 +93,16 @@ module File =
     open System.IO
 
     /// File.ReadAllText fails with locked files. See https://stackoverflow.com/a/1389172
-    let readAllTextNonBlocking (path: string) =
+    let readAllTextNonBlocking (path: string) = async {
         if File.Exists(path) then
             use fileStream = new FileStream(path, FileMode.Open, FileAccess.Read, FileShare.ReadWrite)
             use textReader = new StreamReader(fileStream)
-            textReader.ReadToEnd()
+            let! text = textReader.ReadToEndAsync() |> Async.AwaitTask
+            return text
         else
             Log.always("File does not exist: " + path)
-            ""
-
-    let getRelativePathFromCwd (path: string) =
-        Path.GetRelativePath(Directory.GetCurrentDirectory(), path)
+            return ""
+    }
 
     let rec tryFindPackageJsonDir dir =
         if File.Exists(Path.Combine(dir, "package.json")) then Some dir
@@ -150,6 +149,21 @@ module Process =
     let isWindows() =
         InteropServices.RuntimeInformation.IsOSPlatform(InteropServices.OSPlatform.Windows)
 
+    // Adapted from https://stackoverflow.com/a/22210859
+    let tryFindInPath (exec: string) =
+        let isWindows = isWindows()
+        let exec = if isWindows then exec + ".exe" else exec
+        Environment.GetEnvironmentVariable("PATH")
+            .Split(if isWindows then ';' else ':')
+        |> Array.tryPick (fun dir ->
+            let execPath = IO.Path.Combine(dir, exec)
+            if IO.File.Exists execPath then Some execPath else None)
+
+    let findInPath (exec: string) =
+        match tryFindInPath exec with
+        | Some exec -> exec
+        | None -> failwith $"Cannot find {exec} in PATH"
+
     let getCurrentAssembly() =
         typeof<TypeInThisAssembly>.Assembly
 
@@ -158,20 +172,21 @@ module Process =
         IO.Path.GetFullPath(dir) + (if isWindows() then ";" else ":") + currentPath
 
     // Adapted from https://github.com/enricosada/dotnet-proj-info/blob/1e6d0521f7f333df7eff3148465f7df6191e0201/src/dotnet-proj/Program.fs#L155
-    let private startProcess workingDir exePath args =
-        let args = String.concat " " args
+    let private startProcess (envVars: (string * string) list) workingDir exePath (args: string list) =
         let exePath, args =
-            if isWindows() then "cmd", ("/C " + exePath + " " + args)
+            if isWindows() then "cmd", "/C"::exePath::args
             else exePath, args
 
-        Log.always(File.getRelativePathFromCwd(workingDir) + "> " + exePath + " " + args)
+        // TODO: We should use cliArgs.RootDir instead of Directory.GetCurrentDirectory here but it's only informative
+        // so let's leave it as is for now to avoid having to pass the cliArgs through all the call sites
+        Log.always $"""{IO.Path.GetRelativePath(IO.Directory.GetCurrentDirectory(), workingDir)}> {exePath} {String.concat " " args}"""
 
-        let psi = ProcessStartInfo()
-        // for envVar in envVars do
-        //     psi.EnvironmentVariables.[envVar.Key] <- envVar.Value
-        psi.FileName <- exePath
+        let psi = ProcessStartInfo(exePath)
+        for arg in args do
+            psi.ArgumentList.Add(arg)
+        for (key, value) in envVars do
+            psi.EnvironmentVariables.[key] <- value
         psi.WorkingDirectory <- workingDir
-        psi.Arguments <- args
         psi.CreateNoWindow <- false
         psi.UseShellExecute <- false
 
@@ -182,7 +197,7 @@ module Process =
         if not p.HasExited then
             p.Kill(entireProcessTree=true)
 
-    let start =
+    let startWithEnv envVars =
         let mutable runningProcess = None
 
         // In Windows, terminating the main process doesn't kill the spawned ones so we need
@@ -199,20 +214,26 @@ module Process =
         fun (workingDir: string) (exePath: string) (args: string list) ->
             try
                 runningProcess |> Option.iter kill
-                let p = startProcess workingDir exePath args
+                let p = startProcess envVars workingDir exePath args
                 runningProcess <- Some p
             with ex ->
                 Log.always("Cannot run: " + ex.Message)
 
-    let runSync (workingDir: string) (exePath: string) (args: string list) =
+    let start (workingDir: string) (exePath: string) (args: string list) =
+        startWithEnv [] workingDir exePath args
+
+    let runSyncWithEnv envVars (workingDir: string) (exePath: string) (args: string list) =
         try
-            let p = startProcess workingDir exePath args
+            let p = startProcess envVars workingDir exePath args
             p.WaitForExit()
             p.ExitCode
         with ex ->
             Log.always("Cannot run: " + ex.Message)
             Log.always(ex.StackTrace)
             -1
+
+    let runSync (workingDir: string) (exePath: string) (args: string list) =
+        runSyncWithEnv [] workingDir exePath args
 
 [<RequireQualifiedAccess>]
 module Async =
@@ -342,11 +363,11 @@ module Observable =
                 { new IDisposable with
                     member _.Dispose() = dispose() }
 
-    let throttle ms (obs: IObservable<'T>) =
+    let throttle (ms: int) (obs: IObservable<'T>) =
         { new IObservable<'T[]> with
             member _.Subscribe w =
                 let events = Collections.Concurrent.ConcurrentBag()
-                let timer = new Timers.Timer(ms, AutoReset=false)
+                let timer = new Timers.Timer(float ms, AutoReset=false)
                 timer.Elapsed.Add(fun _ ->
                     events.ToArray() |> w.OnNext
                     timer.Dispose())