+++ src/Fable.Transforms/Replacements.fs
@@ -201,22 +201,25 @@ let (|NumberExt|_|) = function
 let genericTypeInfoError (name: string) =
     $"Cannot get type info of generic parameter {name}. Fable erases generics at runtime, try inlining the functions so generics can be resolved at compile time."
 
-let getTypeNameFromFullName (fullname: string) =
+let splitFulName (fullname: string) =
     let fullname =
         match fullname.IndexOf("[") with
         | -1 -> fullname
         | i -> fullname.[..i - 1]
     match fullname.LastIndexOf(".") with
-    | -1 -> fullname
-    | i -> fullname.Substring(i + 1)
+    | -1 -> "", fullname
+    | i -> fullname.Substring(0, i), fullname.Substring(i + 1)
 
-let getTypeName com (ctx: Context) r t =
+let rec getTypeName com (ctx: Context) r t =
     match t with
     | GenericParam name ->
         genericTypeInfoError name
         |> addError com ctx.InlinePath r
-    | _ -> ()
-    getTypeFullName false t |> getTypeNameFromFullName
+        name
+    | Array elemType ->
+        getTypeName com ctx r elemType + "[]"
+    | _ ->
+        getTypeFullName false t |> splitFulName |> snd
 
 let rec namesof com ctx acc e =
     match acc, e with
@@ -1755,7 +1758,7 @@ let strings (com: ICompiler) (ctx: Context) r t (i: CallInfo) (thisArg: Expr opt
     | "Split", Some c, _ ->
         match args with
         // Optimization
-        | [] -> Helper.InstanceCall(c, "split", t, [makeStrConst ""]) |> Some
+        | [] -> Helper.InstanceCall(c, "split", t, [makeStrConst " "]) |> Some
         | [Value(CharConstant _,_) as separator]
         | [StringConst _ as separator]
         | [Value(NewArray([separator],_),_)] ->
@@ -1767,8 +1770,12 @@ let strings (com: ICompiler) (ctx: Context) r t (i: CallInfo) (thisArg: Expr opt
                 | _ -> Value(NewArray([arg1], String), None)
             let args = [arg1; Value(Null Any, None); arg2]
             Helper.LibCall(com, "String", "split", t, c::args, ?loc=r) |> Some
-        | args ->
-            Helper.LibCall(com, "String", "split", t, args, i.SignatureArgTypes, ?thisArg=thisArg, ?loc=r) |> Some
+        | arg1::args ->
+            let arg1 =
+                match arg1.Type with
+                | Array _ -> arg1
+                | _ -> Value(NewArray([arg1], String), None)
+            Helper.LibCall(com, "String", "split", t, arg1::args, i.SignatureArgTypes, ?thisArg=thisArg, ?loc=r) |> Some
     | "Join", None, _ ->
         let methName =
             match i.SignatureArgTypes with
@@ -3080,7 +3087,7 @@ let types (com: ICompiler) (ctx: Context) r t (i: CallInfo) (thisArg: Expr optio
                     let genMap = List.zip (e.GenericParameters |> List.map (fun p -> p.Name)) genArgs |> Map
                     let comp = if ignoreCase then System.StringComparison.OrdinalIgnoreCase else System.StringComparison.Ordinal
                     e.AllInterfaces |> Seq.tryPick (fun ifc ->
-                        let ifcName = getTypeNameFromFullName ifc.Entity.FullName
+                        let ifcName = splitFulName ifc.Entity.FullName |> snd
                         if ifcName.Equals(name, comp) then
                             let genArgs = ifc.GenericArgs |> List.map (function
                                 | GenericParam name as gen -> Map.tryFind name genMap |> Option.defaultValue gen
@@ -3091,11 +3098,7 @@ let types (com: ICompiler) (ctx: Context) r t (i: CallInfo) (thisArg: Expr optio
                         | Some(ifcEnt, genArgs) -> Value(TypeInfo(DeclaredType(ifcEnt, genArgs)), r)
                         | None -> Value(Null t, r))
             | "get_FullName" -> getTypeFullName false exprType |> returnString r
-            | "get_Namespace" ->
-                let fullname = getTypeFullName false exprType
-                match fullname.LastIndexOf(".") with
-                | -1 -> "" |> returnString r
-                | i -> fullname.Substring(0, i) |> returnString r
+            | "get_Namespace" -> getTypeFullName false exprType |> splitFulName |> fst |> returnString r
             | "get_IsArray" ->
                 match exprType with Array _ -> true | _ -> false
                 |> BoolConstant |> makeValue r |> Some