+++ src/fable-library/Seq.fs
@@ -239,7 +239,7 @@ let ofArray (arr: 'T[]) =
 
 let toArray (xs: seq<'T>): 'T[] =
     match xs with
-    | :? array<'T> as a -> a
+    // | :? array<'T> as a -> Array.ofSeq a
     | :? list<'T> as a -> Array.ofList a
     | _ -> Array.ofSeq xs