+++ src/Fable.Cli/RELEASE_NOTES.md
@@ -1,3 +1,25 @@
+### 3.4.10
+
+* Add support for StringSplitOptions.TrimEntries @steveofficer
+* Fix #2587: DateTimeOffset.Parse issues with some locales @ncave
+* Use Process.ArgumentList to escape args passed to subprocess
+* Print paths relative to --cwd if set
+
+### 3.4.9
+
+* Add CLI arg --watchDelay
+* Show relative paths in logs
+* Fixed Seq.toArray @ncave
+* Fix FullName/Name/Namespace of complex array types
+
+### 3.4.8
+
+* Fix #2572 #2579: Watch .fsi files and referenced .fsproj
+* Fix #2576: Last file can omit module declaration
+* Fix Seq.cache with partial enumeration
+* Fix DateTime Offset parsing with date only and hyphens
+* Set NODE_ENV when running a sub-process
+
 ### 3.4.7
 
 * Fix #2571: Forward slash not escaped when creating regex